% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-as.data.frame.R
\docType{methods}
\name{as.dataframe}
\alias{as.dataframe}
\alias{as.dataframe,Points-method}
\alias{as.dataframe,UncertainPoints-method}
\title{Transformation of S4 class Points or UncertainPoints into data.frame}
\usage{
\S4method{as.dataframe}{UncertainPoints}(data)

\S4method{as.dataframe}{Points}(data)
}
\arguments{
\item{data}{Input data type of S4 object of class UncertainPoints or Points.}
}
\value{
Returns an object of class \code{data.frame}.
}
\description{
This function provides the transformation of S4 object class \code{UncertainPoints} or 
\code{Points} into the \code{data.frame} data format.
}
\seealso{
\code{\link[UncerIn2]{Points-class}}, 
\code{\link[UncerIn2]{Points}}, \code{\link[UncerIn2]{UncertainPoints-class}}, 
\code{\link[UncerIn2]{uncertaintyInterpolation2-package}}
}

