% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-as.UncertainPoints.R
\docType{methods}
\name{as.UncertainPoints}
\alias{as.UncertainPoints}
\alias{as.UncertainPoints,UncertainInterpolation-method}
\title{Transformation of S4 class UncertainInterpolation into UncertainPoints}
\usage{
\S4method{as.UncertainPoints}{UncertainInterpolation}(object,grid)
}
\arguments{
\item{object}{Input data type of S4 object class UncertainInterpolation.}

\item{grid}{Input grid (default set FALSE)}
}
\value{
Returns an object of class \code{UncertainPoints}.
}
\description{
This function provides the transformation of S4 object class \code{UncertainInterpolation} 
into the S4 class \code{UncertainPoints}..
}
\seealso{
\code{\link[UncerIn2]{UncertainInterpolation-class}}, \code{\link[UncerIn2]{UncertainPoints-class}}, \code{\link[UncerIn2]{uncertaintyInterpolation2-package}}
}

