% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-04-FuzzyInterpolation.R
\docType{class}
\name{FuzzyInterpolation-class}
\alias{FuzzyInterpolation-class}
\title{S4 class Representing a FuzzyInterpolation}
\description{
Definition of S4 class for output data from fuzzy (kriging) 
interpolation function of package \code{UncerIn2}. Where 
\code{x, y} should represent coordinates and \code{minimal, 
modalValue, maximal} the values of uncertainty.
}
\section{Slots}{

 \describe{
   \item{\code{x}:}{Input numeric data (number specyfing the x grid coordinate).}
   \item{\code{y}:}{Input numeric data (number specyfing the y grid coordinate).}
   \item{\code{minimal}:}{Defined model values of uncertaintyLower - the bottom calculated part.}
   \item{\code{modalValue}:}{Defined model values of modalValue - the mean values.}
   \item{\code{maximal}:}{Defined model values of uncertaintyUpper - the upper calculated part.}
 }
}
\examples{
showClass("FuzzyInterpolation")
}
\seealso{
\code{\link[UncerIn2]{uncertaintyInterpolation2-package}}

Other res: \code{\link{Points-class}},
  \code{\link{UncertainInterpolation-class}},
  \code{\link{UncertainPoints-class}}
}

