% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Grid.R
\docType{methods}
\name{Grid.box}
\alias{Grid.box}
\alias{Grid.box,Points-method}
\title{Generation of grid defined by surrounding box}
\usage{
\S4method{Grid.box}{Points}(object,gridded,cellsize)
}
\arguments{
\item{object}{Input data type of S4 object class \code{Points}.}

\item{gridded}{Logical value \code{TRUE/FALSE}. If true, then Spatial grid is made.}

\item{cellsize}{Numeric value for the size of cell.}
}
\value{
Returns an object of class \code{SpatialPixels} or \code{data.frame}, depends on whether 
the spatial grid is required.
}
\description{
This method creates a grid defined in two possible ways - the user should decide if spatial grid is 
required (\code{gridded = TRUE/FALSE}). The size of the grid is recognized by the input data coordinates. 
User should define size of the single cell.
}
\seealso{
\code{\link[UncerIn2]{Points-class}}, \code{\link[UncerIn2]{Grid.def}}, \code{\link[UncerIn2]{Grid.interpolation}}, \code{\link[UncerIn2]{uncertaintyInterpolation2-package}}
}

