% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Grid.R
\docType{methods}
\name{Grid.def}
\alias{Grid.def}
\alias{Grid.def,Points-method}
\title{Generation of grid with defined number of cels}
\usage{
\S4method{Grid.def}{Points}(object,gridded,numberOfCellsX,numberOfCellsY)
}
\arguments{
\item{object}{Input data type of S4 object class \code{Points}.}

\item{gridded}{Logical value \code{TRUE/FALSE}. If true, then Spatial grid is made.}

\item{numberOfCellsX}{The number of cells on axle X.}

\item{numberOfCellsY}{The number of cells on axle Y.}
}
\value{
Returns an object of class \code{SpatialPixels} or \code{data.frame}, depends on whether 
the spatial grid is required.
}
\description{
This method creates a grid defined in two possible ways - the user should decide if spatial grid is 
required (\code{gridded = TRUE/FALSE}). The size of the grid should be also defined by the user.
}
\seealso{
\code{\link[UncerIn2]{Points-class}}, \code{\link[UncerIn2]{Grid.box}}, \code{\link[UncerIn2]{Grid.interpolation}}, \code{\link[UncerIn2]{uncertaintyInterpolation2-package}}
}

