% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Grid.R
\docType{methods}
\name{Grid.interpolation}
\alias{Grid.interpolation}
\alias{Grid.interpolation,UncertainInterpolation-method}
\title{Generation of grid from the results of interpolation functions}
\usage{
\S4method{Grid.interpolation}{UncertainInterpolation}(object,gridded,cellsize)
}
\arguments{
\item{object}{Input data type of S4 object class \code{UncertainInterpolation}.}

\item{gridded}{Logical value \code{TRUE/FALSE}. If true, then Spatial grid is made.}

\item{cellsize}{Numeric value for the size of cell.}
}
\value{
Returns an object of class \code{SpatialPixels} or \code{data.frame}, depends on whether 
the spatial grid is required.
}
\description{
This method creates a grid from S4 object class \code{UncertainInterpolation}, defined in two possible 
ways - the user should decide if spatial grid is required (\code{gridded = TRUE/FALSE}). The size of 
the grid is recognized by the input data coordinates. User should define size of the single cell.
}
\seealso{
\code{\link[UncerIn2]{UncertainInterpolation-class}}, \code{\link[UncerIn2]{Grid.box}}, \code{\link[UncerIn2]{Grid.def}}, \code{\link[UncerIn2]{uncertaintyInterpolation2-package}}
}

