% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-idwUncertain.R
\docType{methods}
\name{idwUncertain}
\alias{idwUncertain}
\alias{idwUncertain,UncertainPoints,Spatial-method}
\title{IDW interpolation}
\usage{
\S4method{idwUncertain}{UncertainPoints,Spatial}(object,grid,nmax,idp)
}
\arguments{
\item{object}{Input data. An object of \code{UncertainPoints} class.}

\item{grid}{Input Spatial grid.}

\item{nmax}{The number of nearest observations that should be used.}

\item{idp}{Inverse distance power.}
}
\value{
Returns an object of class \code{UncertainInterpolation}.
}
\description{
This function provides IDW interpolation over the input data enriched by the 
uncertainty. The input data must be an S4 object class of \code{UncertainPoints} 
and grid type of \code{Spatial}. Output object is type of S4 class 
\code{UncertainInterpolation}.
}
\seealso{
\code{\link[UncerIn2]{UncertainPoints-class}}, \code{\link[UncerIn2]{UncertainInterpolation-class}}, \code{\link[UncerIn2]{Grid.def}}, \code{\link[UncerIn2]{Grid.box}}, \code{\link[UncerIn2]{Grid.interpolation}}, \code{\link[gstat]{gstat}}, \code{\link[UncerIn2]{Plot}}, \code{\link[UncerIn2]{uncertaintyInterpolation2-package}}
}

