% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\docType{methods}
\name{Plot}
\alias{Plot}
\alias{Plot,UncertainInterpolation-method}
\title{Plotting S4 class UncertainInterpolation}
\usage{
\S4method{Plot}{UncertainInterpolation}(object, attr1, attr2, attr3, cuts, pretty)

\S4method{Plot}{UncertainInterpolation}(object, attr1 = "uncertaintyLower", attr2 = "modalValue", 
attr3 = "uncertaintyUpper", cuts = 10, pretty=TRUE)
}
\arguments{
\item{object}{Input data type of S4 object class UncertainInterpolation.}

\item{attr1}{First plotting atrribute.}

\item{attr2}{Second plotting atrribute.}

\item{attr3}{Third plotting atrribute.}

\item{cuts}{Number of cuts.}

\item{pretty}{Logical value \code{TRUE/FALSE.}(choose colour breaks at pretty numbers?)}
}
\description{
This function provides the plotting of S4 object class \code{UncertainInterpolation}.
}
\seealso{
\code{\link[UncerIn2]{UncertainInterpolation-class}}, \code{\link[UncerIn2]{uncertaintyInterpolation2-package}}
}

