% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-01-Points.R
\docType{class}
\name{Points-class}
\alias{Points-class}
\title{S4 class Representing a class Points}
\description{
Definition of class for input data of numeric format \code{x/y/z}. 
Also defines input data format into the functions of package \code{UncerIn2}, 
which are building uncertainty models over input data. Where \code{x, y} should 
represent coordinates and \code{z} variable.
}
\section{Slots}{

 \describe{
   \item{\code{x}:}{Input numeric data (number specyfing the x coordinate).}
   \item{\code{y}:}{Input numeric data (number specyfing the y coordinate).}
   \item{\code{z}:}{Input numeric data (number specyfing the values of variable).}
 }
}
\examples{
showClass("Points")
}
\seealso{
\code{\link[UncerIn2]{uncertaintyInterpolation2-package}}

Other res: \code{\link{FuzzyInterpolation-class}},
  \code{\link{UncertainInterpolation-class}},
  \code{\link{UncertainPoints-class}}
}

