% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-01-Points.R
\name{Points}
\alias{Points}
\title{Creates S4 object class Points}
\usage{
Points(x, y, z)
}
\arguments{
\item{x}{Input numeric data (number specyfing the x coordinate).}

\item{y}{Input numeric data (number specyfing the y coordinate).}

\item{z}{Input numeric data (number specyfing the values of variable).}
}
\value{
Returns an object of class \code{Points}.
}
\description{
This function creates an object of S4 class \code{Points}.
}

