% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-03-UncertainInterpolation.R
\docType{class}
\name{UncertainInterpolation-class}
\alias{UncertainInterpolation-class}
\title{S4 class Representing a UncertainInterpolation}
\description{
Definition of S4 class for output data from interpolation 
functions of package \code{UncerIn2}. Where \code{x, y} should 
represent coordinates and \code{uncertaintyLower, modalValue, 
uncertaintyUpper} the values of uncertainty.
}
\section{Slots}{

 \describe{
   \item{\code{x}:}{Input numeric data (number specyfing the x grid coordinate).}
   \item{\code{y}:}{Input numeric data (number specyfing the y grid coordinate).}
   \item{\code{uncertaintyLower}:}{Defined model values of uncertaintyLower - the bottom calculated part.}
   \item{\code{modalValue}:}{Defined model values of modalValue - the mean values.}
   \item{\code{uncertaintyUpper}:}{Defined model values of uncertaintyUpper - the upper calculated part.}
 }
}
\examples{
showClass("UncertainInterpolation")
}
\seealso{
\code{\link[UncerIn2]{uncertaintyInterpolation2-package}}

Other res: \code{\link{FuzzyInterpolation-class}},
  \code{\link{Points-class}},
  \code{\link{UncertainPoints-class}}
}

