% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-uncertaintyConstant.R
\name{uncertaintyConstant}
\alias{uncertaintyConstant}
\title{Creates S4 object class UncertainPoints}
\usage{
uncertaintyConstant(data, value = 2)
}
\arguments{
\item{data}{Input data. S4 class of \code{Points}.}

\item{value}{Constant value of imprecision that defines the uncertainty.}
}
\value{
Returns an object of class \code{UncertainPoints}.
}
\description{
Builds an uncertainty model based on the constant error over the input data. 
Input data must be type of S4 class \code{Points}. Output object 
is type of S4 class \code{UncertainPoints}.
}
\seealso{
\code{\link[UncerIn2]{Points-class}}, \code{\link[UncerIn2]{UncertainPoints-class}}, \code{\link[UncerIn2]{uncertaintyInterpolation2-package}}
}

