% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-uncertaintyPercent.R
\name{uncertaintyPercent}
\alias{uncertaintyPercent}
\title{Creates S4 object class UncertainPoints}
\usage{
uncertaintyPercent(data, value = 2)
}
\arguments{
\item{data}{Input data. S4 class of \code{Points}.}

\item{value}{Percent value of imprecision that defines the uncertainty.}
}
\value{
Returns an object of class \code{UncertainPoints}.
}
\description{
Builds an uncertainty model based on the percentual error 
over the input data. Input data must be type of S4 class 
\code{Points}. Output object is type of S4 class \code{UncertainPoints}.
}
\seealso{
\code{\link[UncerIn2]{Points-class}}, \code{\link[UncerIn2]{UncertainPoints-class}}, \code{\link[UncerIn2]{uncertaintyInterpolation2-package}}
}

