% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-uncertaintyRandomDeviate.R
\name{uncertaintyRandomDeviate}
\alias{uncertaintyRandomDeviate}
\title{Creates S4 object class UncertainPoints}
\usage{
uncertaintyRandomDeviate(data, mean = 0, sd = 1)
}
\arguments{
\item{data}{Input data. S4 class of \code{Points}.}

\item{mean}{The mean value of input data values.}

\item{sd}{The standart deviation of input data values.}
}
\value{
Returns an object of class \code{UncertainPoints}.
}
\description{
Builds an uncertainty model based on the imprecision made by random 
deviates over the input data. Input data must be type of S4 class 
\code{Points}. Output object is type of S4 class \code{UncertainPoints}.
}
\seealso{
\code{\link[UncerIn2]{Points-class}}, \code{\link[UncerIn2]{UncertainPoints-class}}, \code{\link[UncerIn2]{uncertaintyInterpolation2-package}}
}

