% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-uncertaintyRandomNumber.R
\name{uncertaintyRandomNumber}
\alias{uncertaintyRandomNumber}
\title{Creates S4 object class UncertainPoints}
\usage{
uncertaintyRandomNumber(data, min = 0, max = 1)
}
\arguments{
\item{data}{Input data. S4 class of \code{Points}.}

\item{min}{The minimum value of the interval.}

\item{max}{The maximum value of the interval.}
}
\value{
Returns an object of class \code{UncertainPoints}.
}
\description{
Builds an uncertainty model based on the imprecision made by 
random number (from defined interval) over the input data. 
Input data must be type of S4 class \code{Points}. Output 
object is type of S4 class \code{UncertainPoints}.
}
\seealso{
\code{\link[UncerIn2]{Points-class}}, \code{\link[UncerIn2]{UncertainPoints-class}}, \code{\link[UncerIn2]{uncertaintyInterpolation2-package}}
}

